Move model-asset helpers to helpers/modelAssets.ts
Relocate modelAssetIdForFeature, normalizeImportedModelStorage,
pruneUnusedModelAssets, and isImportedModelFeature into a dedicated
home module. Slices now import these directly.
